requirements.txt
asgiref==3.11.0
crispy-bootstrap4==2025.6
dj-database-url==3.0.1
Django==5.2.9
django-configurations==2.5.1
django-crispy-forms==2.5
sqlparse==0.5.5
typing_extensions==4.15.0
